move common scripts to eng/common
