+++ eng/common/scripts/helpers.js
