+++ eng/tsp-core/scripts/upload-bundler-packages.js
@@ -3,7 +3,7 @@ import {
   bundleAndUploadPackages,
   getPackageVersion,
 } from "../../../packages/bundle-uploader/dist/src/index.js";
-import { repoRoot } from "./helpers.js";
+import { repoRoot } from "../../common/scripts/helpers.js";
 
 await bundleAndUploadPackages({
   repoRoot: repoRoot,
